pyproject.toml
[project]
name = "slcm-scraper"
version = "0.1.0"
description = "CGPA scraper for Manipal SLCM portal"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.33.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
